requirements.txt
django~=5.1.1
psycopg2
djangorestframework
drf-yasg
django-environ
environs
pillow
setuptools